requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
dataclasses-json==0.5.14
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
langchain==0.0.288
langsmith==0.0.37
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.24.4
openai==0.28.0
packaging==23.1
phospho==0.0.7
pydantic==2.3.0
pydantic-core==2.6.3
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.20
tenacity==8.2.3
tqdm==4.66.1
typing-extensions==4.7.1
typing-inspect==0.9.0
urllib3==2.0.4
yarl==1.9.2
